requirements.txt
lxml 
requests 
jieba 
numpy 
Pillow
openpyxl 
wordcloud 
pyecharts